+++ package.json
@@ -7,8 +7,12 @@
     "deploy": "gh-pages -d dist",
     "watch": "webpack --watch",
     "start": "webpack serve --open",
-    "build": "webpack"
+    "build": "webpack",
+    "test": "jest --coverage"
   },
+  "jest": {
+    "testEnvironment": "jsdom"
+},
   "repository": {
     "type": "git",
     "url": "git+https://github.com/naji-97/To-Do-list.git"
@@ -21,6 +25,7 @@
   },
   "homepage": "https://github.com/naji-97/To-Do-list",
   "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.20.11",
     "babel-eslint": "^10.1.0",
     "css-loader": "^6.7.3",
     "eslint": "^7.32.0",
@@ -28,6 +33,7 @@
     "eslint-plugin-import": "^2.27.5",
     "hint": "^7.1.3",
     "html-webpack-plugin": "^5.5.0",
+    "jest": "^29.4.1",
     "style-loader": "^3.3.1",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
@@ -38,7 +44,9 @@
     "webpack-dev-server": "^4.11.1"
   },
   "dependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
     "gh-pages": "^5.0.0",
+    "jest-environment-jsdom": "^29.4.1",
     "lodash": "^4.17.21"
   }
 }
